feat: Mejoras funcionales y de estilos fix: Max degree method. Falta un control de errores

+++ src/main/java/model/Form2.java
@@ -44,7 +44,7 @@ public class Form2  extends Polynomial<Form2, int[]> {
     }
 
     @Override
-    public double eval(int x) {
+    public int eval(int x) {
         return 0;
     }
